Various bug fixes and QoL improvements
